add MapContainer and StatusMessage components
All data needed by the Map component is loaded in the MapContainer and
then passed as 'geoJSON' and 'latlng' props to the Map component.

The 'status' state object in the MapContainer keeps track of loading
progress and any errors that may occur durring this process and is
passed as a prop to the StatusMessage component for display to the user.