remove map.locate() call from componentDidMount in Map component
We are using the browser's geolocation API to handle retrieval of
the user's position in the MapContainer and passing this data to the
Map component, so we don't need to use Leaflet's location API here.

Also updated the shape of the 'pos' object passed to handleLocationFound
if onLocationfound event is emitted so that it fits the shape that the
handler is expecting (the pos object returned from Leaflet's location event
is slightly different than returned from Geolocation.getCurrentPosition).